Add vintage parser and helper functions
Adds parse_single vintage, as well as several helper functions.
This allows to read vintages passed from previous data cleaning steps,
 e.g., 1960-1961-1, and interpret them as ranges of integers (1946, 1960).

For cases with incomplete years, marked by the final '-1' value, the label
is split in two parts: one for the complete years, in this case, 1960-1960,
and one for the incomplete years, 1961-1965.

A share is also returned, based on the total number of months with construction
activity in the split vintage. By default, we assume a linear distribution (i.e., 1/12)
of construction activity over any given year.