Harmonize census vintages while ensuring data accuracy
Add functions, helper functions and related unit tests to harmonize
the different census datasets.

The main function, harmonize_vintage_labels, goes over each dataframe,
row by row. Then, it parses the vintage, and splits it into several
vintages if required.

There are checks to validate that the shares are respected, as well as
the sums.